3791 - address other review comments
