+++ barretenberg/cpp/src/barretenberg/vm/avm_trace/AvmMini_execution.hpp
@@ -15,8 +15,10 @@ class Execution {
     Execution() = default;
 
     static size_t const AVM_OPERAND_BYTE_LENGTH = 4; // Keep in sync with TS code
-    static size_t const AVM_OPCODE_BYTE_LENGTH = 1;  // Keep in sync with TS code
-    static size_t const AVM_IN_TAG_BYTE_LENGTH = 1;  // Keep in sync with TS code
+    static_assert(sizeof(uint32_t) / sizeof(uint8_t) == AVM_OPERAND_BYTE_LENGTH);
+
+    static size_t const AVM_OPCODE_BYTE_LENGTH = 1; // Keep in sync with TS code
+    static size_t const AVM_IN_TAG_BYTE_LENGTH = 1; // Keep in sync with TS code
 
     static std::vector<Instruction> parse(std::vector<uint8_t> const& bytecode);
     static std::vector<Row> gen_trace(std::vector<Instruction> const& instructions, std::vector<FF> const& calldata);
